+++ compiler-core/src/type_/error.rs
@@ -925,6 +925,7 @@ pub enum FeatureKind {
     RecordAccessVariantInference,
     LetAssertWithMessage,
     VariantWithDeprecatedAnnotation,
+    JavaScriptUnalignedBitArray,
 }
 
 impl FeatureKind {
@@ -950,6 +951,8 @@ impl FeatureKind {
             FeatureKind::VariantWithDeprecatedAnnotation | FeatureKind::LetAssertWithMessage => {
                 Version::new(1, 7, 0)
             }
+
+            FeatureKind::JavaScriptUnalignedBitArray => Version::new(1, 9, 0),
         }
     }
 }