+++ compiler-core/src/type_/expression.rs
@@ -1172,6 +1172,40 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             }
         })?;
 
+        // Track usage of the unaligned bit arrays feature on JavaScript so that
+        // warnings can be emitted if the Gleam version constraint is too low
+        if self.environment.target == Target::JavaScript
+            && !self.current_function_definition.has_javascript_external
+        {
+            for option in options.iter() {
+                if let BitArrayOption::<TypedValue>::Size {
+                    value, location, ..
+                } = option
+                {
+                    let mut using_unaligned_bit_array = false;
+
+                    if type_ == int() {
+                        match &(**value).as_int_literal() {
+                            Some(size) if size % 8 != 0 => {
+                                using_unaligned_bit_array = true;
+                            }
+                            _ => (),
+                        }
+                    } else if type_ == bits() {
+                        using_unaligned_bit_array = true;
+                    }
+
+                    if using_unaligned_bit_array {
+                        self.track_feature_usage(
+                            FeatureKind::JavaScriptUnalignedBitArray,
+                            *location,
+                        );
+                        break;
+                    }
+                }
+            }
+        }
+
         unify(type_.clone(), value.type_())
             .map_err(|e| convert_unify_error(e, value.location()))?;
 
@@ -1332,6 +1366,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         let mut pattern_typer = pattern::PatternTyper::new(
             self.environment,
             &self.implementations,
+            &self.current_function_definition,
             &self.hydrator,
             self.problems,
         );
@@ -1606,6 +1641,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         let mut pattern_typer = pattern::PatternTyper::new(
             self.environment,
             &self.implementations,
+            &self.current_function_definition,
             &self.hydrator,
             self.problems,
         );