+++ compiler-core/src/type_/pattern.rs
@@ -1,6 +1,7 @@
 use hexpm::version::Version;
 use im::hashmap;
 use itertools::Itertools;
+use num_bigint::BigInt;
 
 /// Type inference and checking of patterns used in case expressions
 /// and variables bindings.
@@ -11,12 +12,14 @@ use crate::{
     ast::{
         AssignName, BitArrayOption, ImplicitCallArgOrigin, Layer, UntypedPatternBitArraySegment,
     },
+    type_::expression::FunctionDefinition,
 };
 use std::sync::Arc;
 
 pub struct PatternTyper<'a, 'b> {
     environment: &'a mut Environment<'b>,
     implementations: &'a Implementations,
+    current_function: &'a FunctionDefinition,
     hydrator: &'a Hydrator,
     mode: PatternMode,
     initial_pattern_vars: HashSet<EcoString>,
@@ -38,12 +41,14 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
     pub fn new(
         environment: &'a mut Environment<'b>,
         implementations: &'a Implementations,
+        current_function: &'a FunctionDefinition,
         hydrator: &'a Hydrator,
         problems: &'a mut Problems,
     ) -> Self {
         Self {
             environment,
             implementations,
+            current_function,
             hydrator,
             mode: PatternMode::Initial,
             initial_pattern_vars: HashSet::new(),
@@ -296,6 +301,43 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                 location: error.location,
             })?;
 
+        // Track usage of the unaligned bit arrays feature on JavaScript so that
+        // warnings can be emitted if the Gleam version constraint is too low
+        if self.environment.target == Target::JavaScript
+            && !self.current_function.has_javascript_external
+        {
+            for option in options.iter() {
+                match option {
+                    // Use of the `bits` segment type
+                    BitArrayOption::<TypedPattern>::Bits { location } => {
+                        self.track_feature_usage(
+                            FeatureKind::JavaScriptUnalignedBitArray,
+                            *location,
+                        );
+                    }
+
+                    // Int segments that aren't a whole number of bytes
+                    BitArrayOption::<TypedPattern>::Size { value, .. } if segment_type == int() => {
+                        match &**value {
+                            Pattern::<_>::Int {
+                                location,
+                                int_value,
+                                ..
+                            } if int_value % 8 != BigInt::ZERO => {
+                                self.track_feature_usage(
+                                    FeatureKind::JavaScriptUnalignedBitArray,
+                                    *location,
+                                );
+                            }
+                            _ => (),
+                        }
+                    }
+
+                    _ => (),
+                }
+            }
+        }
+
         let type_ = {
             match value.deref() {
                 Pattern::Variable { .. } if segment_type == string() => {
@@ -472,7 +514,7 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                 unify(type_, int()).map_err(|e| convert_unify_error(e, location))?;
 
                 if self.environment.target == Target::JavaScript
-                    && !self.implementations.uses_javascript_externals
+                    && !self.current_function.has_javascript_external
                 {
                     check_javascript_int_safety(&int_value, location, self.problems);
                 }