+++ test/javascript_prelude/main.mjs
@@ -13,6 +13,7 @@ import {
   toBitArray,
   toList,
   sizedInt,
+  sizedFloat,
 } from "./prelude.mjs";
 
 let failures = 0;
@@ -134,20 +135,56 @@ assertNotEqual(List.fromArray([1, 2]), List.fromArray([1, 2, new Ok(2)]));
 assertNotEqual(List.fromArray([1]), List.fromArray([]));
 assertNotEqual(List.fromArray([]), List.fromArray([1]));
 
-assertEqual(new BitArray(new Uint8Array([])), new BitArray(new Uint8Array([])));
-assertEqual(
-  new BitArray(new Uint8Array([1, 2, 3])),
-  new BitArray(new Uint8Array([1, 2, 3])),
-);
-assertNotEqual(
-  new BitArray(new Uint8Array([1, 2])),
-  new BitArray(new Uint8Array([1, 2, 3])),
-);
-
 assertEqual(new UtfCodepoint(128013), new UtfCodepoint(128013));
 assertNotEqual(new UtfCodepoint(128013), new UtfCodepoint(128014));
 
-// toBitArray
+// new BitArray()
+
+assertEqual(
+  new BitArray(new Uint8Array([1, 2, 3])),
+  new BitArray(new Uint8Array([1, 2, 3]), 24),
+);
+
+assertThrows("`new BitArray()` throws with an ArrayBuffer", () => {
+  new BitArray(new ArrayBuffer(8));
+});
+
+assertThrows(
+  "`new BitArray()` throws with a raw array",
+  () => new BitArray([1, 2]),
+);
+
+assertThrows(
+  "`new BitArray()` throws with invalid bit size",
+  () => new BitArray(new Uint8Array([2]), -1),
+);
+
+assertThrows(
+  "`new BitArray()` throws with too many bytes for the bit size",
+  () => new BitArray(new Uint8Array([1, 2]), 7),
+);
+
+assertThrows(
+  "`new BitArray()` throws with too few bytes for the bit size",
+  () => new BitArray(new Uint8Array([1, 2]), 17),
+);
+
+assertThrows(
+  "`new BitArray()` throws with an invalid bit offset",
+  () => new BitArray(new Uint8Array([1]), 0, -1),
+);
+
+assertThrows(
+  "`new BitArray()` throws with an invalid bit offset",
+  () => new BitArray(new Uint8Array([1]), 0, 8),
+);
+
+// toBitArray()
+
+assertEqual(
+  new BitArray(new Uint8Array([1, 2])),
+  toBitArray([new BitArray(new Uint8Array([1, 2]))]),
+);
 
 assertEqual(new BitArray(new Uint8Array([])), toBitArray([]));
 
@@ -179,17 +216,17 @@ for (const { input, u8 } of testValues) {
 assertEqual(new BitArray(new Uint8Array([])), toBitArray([new Uint8Array([])]));
 assertEqual(
   new BitArray(new Uint8Array([1, 2, 4, 8])),
-  toBitArray([new Uint8Array([1, 2, 4, 8])])
+  toBitArray([new Uint8Array([1, 2, 4, 8])]),
 );
 
 assertEqual(
   new BitArray(new Uint8Array(testValues.map((t) => t.u8))),
-  toBitArray(testValues.map((t) => t.input))
+  toBitArray(testValues.map((t) => t.input)),
 );
 
 assertEqual(
   new BitArray(
-    new Uint8Array([1, 2, 4, 8, ...testValues.map((t) => t.u8), 80, 90, 100])
+    new Uint8Array([1, 2, 4, 8, ...testValues.map((t) => t.u8), 80, 90, 100]),
   ),
   toBitArray([
     new Uint8Array([]),
@@ -198,7 +235,7 @@ assertEqual(
     new Uint8Array([80, 90]),
     new Uint8Array([]),
     new Uint8Array([100]),
-  ])
+  ]),
 );
 
 assertEqual(
@@ -214,6 +251,103 @@ assertEqual(
 assertEqual(
   new BitArray(new Uint8Array([240, 159, 144, 141])),
   toBitArray([codepointBits(new UtfCodepoint(128013))]),
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([240, 159, 144, 141, 0xfe]), 39),
+  toBitArray([
+    new BitArray(new Uint8Array([240, 159, 144])),
+    new BitArray(new Uint8Array([141])),
+    new BitArray(new Uint8Array([0xfe]), 7),
+  ]),
+);
+
+assertEqual(
+  new BitArray(
+    new Uint8Array([240, 159, 144, 0xa9, 0b11101010, 0xf7, 0x39, 0xae]),
+    64,
+  ),
+  toBitArray([
+    new BitArray(new Uint8Array([240, 159, 144])),
+    new BitArray(new Uint8Array([0b10110101]), 4, 3),
+    new BitArray(new Uint8Array([0x9f]), 7),
+    new BitArray(new Uint8Array([0b00010100]), 4, 2),
+    new BitArray(new Uint8Array([0]), 1),
+    new BitArray(new Uint8Array([0xaf, 0x73, 0x9a, 0xee]), 24, 4),
+  ]),
+);
+
+assertEqual(
+  new BitArray(
+    new Uint8Array([
+      129, 145, 57, 255, 255, 255, 255, 255, 191, 157, 243, 182, 246, 62, 104,
+      49, 62, 31, 110, 200, 120, 13, 88,
+    ]),
+    181,
+  ),
+  toBitArray([
+    sizedInt(2, 2, false),
+    sizedInt(0, 2, false),
+    sizedInt(200, 11, true),
+    sizedInt(-100, 49, false),
+    sizedFloat(-1.234, 32, true),
+    sizedFloat(-8.2e40, 64, false),
+    sizedInt(0xf, 5, false),
+    new Uint8Array([1]),
+    0xab,
+  ]),
+);
+
+// BitArray.equals()
+
+assertEqual(new BitArray(new Uint8Array([])), new BitArray(new Uint8Array([])));
+assertEqual(
+  new BitArray(new Uint8Array([1, 2, 3])),
+  new BitArray(new Uint8Array([1, 2, 3])),
+);
+assertNotEqual(
+  new BitArray(new Uint8Array([1, 2])),
+  new BitArray(new Uint8Array([1, 2, 3])),
+);
+assertNotEqual(
+  new BitArray(new Uint8Array([1, 0xf0]), 12),
+  new BitArray(new Uint8Array([2, 0xf0]), 12),
+);
+assertNotEqual(
+  new BitArray(new Uint8Array([1, 0xf0]), 12),
+  new BitArray(new Uint8Array([1, 0xf0]), 13),
+);
+assertNotEqual(
+  new BitArray(new Uint8Array([0x12, 0x30]), 12),
+  new BitArray(new Uint8Array([0x12, 0x4f]), 12),
+);
+assertEqual(
+  new BitArray(new Uint8Array([0x12, 0x30]), 12),
+  new BitArray(new Uint8Array([0x12, 0x3f]), 12),
+);
+assertEqual(
+  new BitArray(new Uint8Array([0b10110110, 0b01101001]), 8, 3),
+  new BitArray(new Uint8Array([0b10110011]), 8),
+);
+assertEqual(
+  new BitArray(new Uint8Array([0b10110110, 0b01101001, 0b10011010]), 17, 5),
+  new BitArray(new Uint8Array([0b11001101, 0b00110011, 0b01000001]), 17),
+);
+assertEqual(
+  new BitArray(new Uint8Array([0b11001101, 0b00110011]), 14, 1),
+  new BitArray(new Uint8Array([0b11100110, 0b10011001]), 14, 2),
+);
+assertEqual(
+  new BitArray(new Uint8Array([0b10110110]), 4, 2),
+  new BitArray(new Uint8Array([0b10111011]), 4, 3),
+);
+assertNotEqual(
+  new BitArray(new Uint8Array([0b10110110, 0b10110110]), 9, 2),
+  new BitArray(new Uint8Array([0b10110111, 0b10100110]), 9, 2),
+);
+assertNotEqual(
+  new BitArray(new Uint8Array([0b10110110, 0b10110110]), 9, 2),
+  new BitArray(new Uint8Array([0b10110110, 0b10000110]), 9, 2),
 );
 
 // toList
@@ -399,77 +533,142 @@ assertNotEqual(hasEqualsField, hasEqualsField2);
 
 assertEqual(new BitArray(new Uint8Array([1, 2, 3])).byteAt(0), 1);
 assertEqual(new BitArray(new Uint8Array([1, 2, 3])).byteAt(2), 3);
-assertEqual(new BitArray(new Uint8Array([1, 2, 3])).intFromSlice(0, 1, true, false), 1);
-assertEqual(new BitArray(new Uint8Array([160, 2, 3])).intFromSlice(0, 1, false, true), -96);
-assertEqual(new BitArray(new Uint8Array([1, 2, 3])).intFromSlice(0, 2, true, false), 258);
-assertEqual(new BitArray(new Uint8Array([1, 2, 3])).intFromSlice(0, 2, false, false), 513);
-assertEqual(new BitArray(new Uint8Array([1, 160, 3])).intFromSlice(0, 2, false, true), -24575);
-assertEqual(new BitArray(new Uint8Array([160, 2, 3])).intFromSlice(0, 2, true, false), 40962);
-assertEqual(new BitArray(new Uint8Array([160, 2, 3])).intFromSlice(0, 2, true, true), -24574);
-assertEqual(
-  new BitArray(new Uint8Array([255, 255, 255, 255, 255, 255, 255])).intFromSlice(0, 7, true, true),
-  -1,
+
+// BitArray.slice()
+
+assertThrows("`BitArray.slice()` throws if start is less than zero", () =>
+  new BitArray(new Uint8Array([1])).slice(-1, 8),
 );
-assertEqual(
-  new BitArray(new Uint8Array([255, 255, 255, 255, 255, 255, 254])).intFromSlice(0, 7, true, false),
-  Number(0xFFFFFFFFFFFFFEn),
+
+assertThrows(
+  "`BitArray.slice()` throws if start is greater than the bit size",
+  () => new BitArray(new Uint8Array([1])).slice(9, 8),
 );
-assertEqual(
-  new BitArray(new Uint8Array([63, 240, 0, 0, 0, 0, 0, 0])).floatFromSlice(0, 8, true),
-  1.0,
+
+assertThrows("`BitArray.slice()` throws if end is before start", () =>
+  new BitArray(new Uint8Array([1])).slice(4, 2),
 );
-assertEqual(
-  new BitArray(new Uint8Array([0, 0, 0, 0, 0, 0, 240, 63])).floatFromSlice(0, 8, false),
-  1.0,
+
+assertThrows(
+  "`BitArray.slice()` throws if end is greater than the bit size",
+  () => new BitArray(new Uint8Array([1])).slice(0, 10),
 );
+
 assertEqual(
-  new BitArray(new Uint8Array([0xC9, 0x74, 0x24, 0x00])).floatFromSlice(0, 4, true),
-  -1000000.0,
+  new BitArray(new Uint8Array([1, 2, 3])).slice(0, 0),
+  new BitArray(new Uint8Array([])),
 );
+
 assertEqual(
-  new BitArray(new Uint8Array([0x00, 0x24, 0x74, 0xC9])).floatFromSlice(0, 4, false),
-  -1000000.0,
+  new BitArray(new Uint8Array([0xbe, 0xff])).slice(0, 2),
+  new BitArray(new Uint8Array([0xbe]), 2),
 );
+
 assertEqual(
-  new BitArray(new Uint8Array([1, 2, 3, 4, 5])).binaryFromSlice(1, 4),
-  new BitArray(new Uint8Array([2, 3, 4])),
+  new BitArray(new Uint8Array([0x12, 0b10101101, 0xff])).slice(10, 14),
+  new BitArray(new Uint8Array([0b10110100]), 4),
 );
+
 assertEqual(
-  new BitArray(new Uint8Array([1, 2, 3])).sliceAfter(1),
+  new BitArray(new Uint8Array([1, 2, 3])).slice(8, 24),
   new BitArray(new Uint8Array([2, 3])),
 );
+
 assertEqual(
-  new BitArray(new Uint8Array([1, 2, 3, 4, 5]))
-    .binaryFromSlice(1, 4)
-    .sliceAfter(1),
-  new BitArray(new Uint8Array([3, 4]))
+  new BitArray(new Uint8Array([1, 2, 3, 4, 5])).slice(8, 32),
+  new BitArray(new Uint8Array([2, 3, 4])),
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([1, 0xfe, 0xa1])).slice(8, 19),
+  new BitArray(new Uint8Array([0xfe, 0xa1]), 11),
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([17, 79, 190, 151, 98, 222, 101])).slice(19, 51),
+  new BitArray(new Uint8Array([244, 187, 22, 243]), 32),
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([0, 0, 0, 0, 0xff, 0xe0])).slice(37, 41),
+  new BitArray(new Uint8Array([0b11111100]), 4),
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([0xaa, 0xbb, 0xff, 0, 0, 0]))
+    .slice(8, 48)
+    .sliceToFloat(8, 40, false),
+  3.5733110840282835e-43,
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([1, 2, 3, 4, 5])).slice(8, 32).slice(8),
+  new BitArray(new Uint8Array([3, 4])),
+);
+
+assertEqual(
+  new BitArray(
+    new Uint8Array([0b00000001, 0b00000010, 0b00000011]),
+    20,
+    2,
+  ).slice(4, 10),
+  new BitArray(new Uint8Array([0b00000001, 0b00000010]), 6, 6),
+);
+
+assertEqual(
+  new BitArray(new Uint8Array([1])).slice(1),
+  new BitArray(new Uint8Array([0b00000010]), 7),
+);
+
+// sizedFloat()
+
+assertEqual(sizedFloat(16.25, 32, true), new Uint8Array([65, 130, 0, 0]));
+assertEqual(sizedFloat(-16.25, 32, false), new Uint8Array([0, 0, 130, 193]));
+assertEqual(
+  sizedFloat(1000.5, 64, true),
+  new Uint8Array([64, 143, 68, 0, 0, 0, 0, 0]),
+);
+assertEqual(
+  sizedFloat(-1000.5, 64, false),
+  new Uint8Array([0, 0, 0, 0, 0, 68, 143, 192]),
 );
 
 // sizedInt()
 
+assertEqual(sizedInt(100, -8, true), new Uint8Array([]));
+assertEqual(sizedInt(100, 0, true), new Uint8Array([]));
+assertEqual(sizedInt(100, 8, true), new Uint8Array([100]));
+assertEqual(sizedInt(-10, 8, true), new Uint8Array([246]));
+
+assertEqual(sizedInt(1, 1, true), new BitArray(new Uint8Array([0x80]), 1));
+assertEqual(sizedInt(2, 2, true), new BitArray(new Uint8Array([0x80]), 2));
+assertEqual(sizedInt(15, 4, true), new BitArray(new Uint8Array([0xf0]), 4));
+assertEqual(sizedInt(100, 7, true), new BitArray(new Uint8Array([200]), 7));
+assertEqual(sizedInt(-44, 7, true), new BitArray(new Uint8Array([168]), 7));
+assertEqual(sizedInt(-1, 6, true), new BitArray(new Uint8Array([252]), 6));
+assertEqual(sizedInt(-1231, 3, true), new BitArray(new Uint8Array([32]), 3));
+assertEqual(sizedInt(1231, 5, true), new BitArray(new Uint8Array([120]), 5));
+
 assertEqual(
-  sizedInt(100, 0, true),
-  new Uint8Array([]),
+  sizedInt(773, 10, true),
+  new BitArray(new Uint8Array([193, 64]), 10),
 );
 assertEqual(
-  sizedInt(0, 32, true),
-  new Uint8Array([0, 0, 0, 0]),
+  sizedInt(-276, 10, false),
+  new BitArray(new Uint8Array([236, 128]), 10),
 );
+assertEqual(sizedInt(80000, 16, true), new Uint8Array([56, 128]));
+assertEqual(sizedInt(-80000, 16, true), new Uint8Array([199, 128]));
+assertEqual(sizedInt(1, 24, true), new Uint8Array([0, 0, 1]));
 assertEqual(
-  sizedInt(1, 24, true),
-  new Uint8Array([0, 0, 1]),
+  sizedInt(-100, 31, false),
+  new BitArray(new Uint8Array([156, 255, 255, 254]), 31),
 );
+assertEqual(sizedInt(0, 32, true), new Uint8Array([0, 0, 0, 0]));
+assertEqual(sizedInt(-1, 32, true), new Uint8Array([255, 255, 255, 255]));
 assertEqual(
-  sizedInt(-1, 32, true),
-  new Uint8Array([255, 255, 255, 255]),
-);
-assertEqual(
-  sizedInt(80000, 16, true),
-  new Uint8Array([56, 128]),
-);
-assertEqual(
-  sizedInt(-80000, 16, true),
-  new Uint8Array([199, 128]),
+  sizedInt(-10, 33, true),
+  new BitArray(new Uint8Array([255, 255, 255, 251, 0]), 33),
 );
 assertEqual(
   sizedInt(-489_391_639_457_909_760, 56, true),
@@ -486,6 +685,255 @@ assertEqual(
 assertEqual(
   sizedInt(Number.MIN_SAFE_INTEGER, 64, true),
   new Uint8Array([255, 224, 0, 0, 0, 0, 0, 1]),
+);
+assertEqual(
+  sizedInt(Number.MAX_SAFE_INTEGER, 77, true),
+  new BitArray(
+    new Uint8Array([0, 0, 0, 255, 255, 255, 255, 255, 255, 248]),
+    77,
+  ),
+);
+assertEqual(
+  sizedInt(Number.MIN_SAFE_INTEGER, 75, false),
+  new BitArray(new Uint8Array([1, 0, 0, 0, 0, 0, 224, 255, 255, 224]), 75),
+);
+assertEqual(
+  sizedInt(Number(9444732965739289353650176n), 75, true),
+  new BitArray(new Uint8Array([255, 255, 255, 255, 255, 248, 0, 0, 0, 0]), 75),
+);
+
+// BitArray.sliceToFloat()
+
+assertEqual(
+  toBitArray([63, 240, 0, 0, 0, 0, 0, 0]).sliceToFloat(0, 64, true),
+  1.0,
+);
+assertEqual(
+  toBitArray([0, 0, 0, 0, 0, 0, 240, 63]).sliceToFloat(0, 64, false),
+  1.0,
+);
+assertEqual(
+  toBitArray([0xff, 0xc9, 0x74, 0x24, 0x00]).sliceToFloat(8, 40, true),
+  -1000000.0,
+);
+assertEqual(
+  toBitArray([0x00, 0x24, 0x74, 0xc9]).sliceToFloat(0, 32, false),
+  -1000000.0,
+);
+assertEqual(
+  new BitArray(
+    new Uint8Array([112, 152, 127, 244, 0, 7, 192]),
+    50,
+    0,
+  ).sliceToFloat(11, 43, true),
+  -511.25,
+);
+assertEqual(
+  new BitArray(
+    new Uint8Array([8, 0, 0, 0, 1, 129, 39, 103, 129, 254]),
+    79,
+    0,
+  ).sliceToFloat(7, 71, false),
+  -5011.75,
+);
+assertEqual(
+  new BitArray(
+    new Uint8Array([212, 152, 127, 244, 0, 7, 192]),
+    50,
+    3,
+  ).sliceToFloat(11, 43, true),
+  1.0714967429001753e-19,
+);
+
+// BitArray.sliceToInt()
+
+assertEqual(toBitArray([0b10011110]).sliceToInt(0, 4, true, false), 0b1001);
+assertEqual(
+  new BitArray(new Uint8Array([0b11001110]), 7, 1).sliceToInt(
+    0,
+    4,
+    true,
+    false,
+  ),
+  0b1001,
+);
+assertEqual(toBitArray([0b10011110]).sliceToInt(4, 8, true, false), 0b1110);
+assertEqual(toBitArray([0b10011110]).sliceToInt(1, 6, true, true), 0b00111);
+assertEqual(toBitArray([0b10011010]).sliceToInt(3, 8, true, true), -6);
+assertEqual(toBitArray([0b10011010]).sliceToInt(0, 7, true, true), -51);
+assertEqual(
+  new BitArray(new Uint8Array([0b11001100, 0b00100000]), 11).sliceToInt(
+    1,
+    11,
+    false,
+    false,
+  ),
+  408,
+);
+assertEqual(toBitArray([0xb6, 0xe3]).sliceToInt(0, 12, true, false), 0xb6e);
+assertEqual(toBitArray([0xb6, 0xe3]).sliceToInt(0, 12, false, false), 0xeb6);
+assertEqual(
+  toBitArray([0xff, 0xb6, 0xe3]).sliceToInt(8, 20, true, false),
+  0xb6e,
+);
+assertEqual(
+  toBitArray([0xff, 0xb6, 0xe3]).sliceToInt(20, 24, true, false),
+  0x03,
+);
+assertEqual(
+  toBitArray([0xff, 0xb6, 0xe3]).sliceToInt(8, 20, false, false),
+  0xeb6,
+);
+assertEqual(
+  toBitArray([0xa5, 0x6c, 0xaa]).sliceToInt(5, 18, true, false),
+  5554,
+);
+assertEqual(
+  toBitArray([0xa5, 0x6c, 0xaa]).sliceToInt(5, 18, false, true),
+  -3411,
+);
+assertEqual(toBitArray([1, 2, 3]).sliceToInt(0, 8, true, false), 1);
+assertEqual(toBitArray([160, 2, 3]).sliceToInt(0, 8, false, true), -96);
+assertEqual(toBitArray([1, 2, 3]).sliceToInt(0, 16, true, false), 258);
+assertEqual(toBitArray([1, 2, 3]).sliceToInt(0, 16, false, false), 513);
+assertEqual(toBitArray([1, 160, 3]).sliceToInt(0, 16, false, true), -24575);
+assertEqual(toBitArray([160, 2, 3]).sliceToInt(0, 16, true, false), 40962);
+assertEqual(toBitArray([3, 160, 2]).sliceToInt(8, 24, true, true), -24574);
+assertEqual(
+  new BitArray(new Uint8Array([146, 192, 70, 25, 128]), 33).sliceToInt(
+    1,
+    24,
+    true,
+    true,
+  ),
+  1_228_870,
+);
+assertEqual(
+  toBitArray([255, 255, 255, 255]).sliceToInt(0, 32, false, true),
+  -1,
+);
+assertEqual(
+  new BitArray(new Uint8Array([217, 150, 209, 191, 0]), 33).sliceToInt(
+    1,
+    33,
+    true,
+    false,
+  ),
+  3_006_112_638,
+);
+assertEqual(
+  new BitArray(new Uint8Array([146, 192, 70, 25, 128]), 33).sliceToInt(
+    1,
+    33,
+    true,
+    true,
+  ),
+  629_181_491,
+);
+assertEqual(
+  new BitArray(new Uint8Array([251, 24, 47, 227, 128]), 33).sliceToInt(
+    1,
+    33,
+    false,
+    false,
+  ),
+  3_344_904_438,
+);
+assertEqual(
+  new BitArray(new Uint8Array([240, 102, 91, 101, 128]), 33).sliceToInt(
+    0,
+    33,
+    false,
+    false,
+  ),
+  5_995_456_240,
+);
+assertEqual(
+  toBitArray([231, 255, 255, 255, 254, 123]).sliceToInt(0, 40, true, true),
+  -103_079_215_106,
+);
+assertEqual(
+  toBitArray([0, 231, 255, 255, 253, 123, 17]).sliceToInt(1, 55, true, false),
+  127_543_348_739_464,
+);
+assertEqual(
+  toBitArray([142, 231, 255, 255, 253, 123, 17, 139]).sliceToInt(
+    8,
+    62,
+    false,
+    true,
+  ),
+  -8425025061257241,
+);
+assertEqual(
+  toBitArray([142, 231, 255, 255, 253, 123, 17, 139]).sliceToInt(
+    7,
+    62,
+    false,
+    true,
+  ),
+  -8293899692933261,
+);
+assertEqual(
+  toBitArray([142, 231, 255, 255, 253, 123, 17]).sliceToInt(8, 48, true, true),
+  -103_079_215_749,
+);
+assertEqual(
+  toBitArray([255, 255, 255, 255, 255, 255, 255]).sliceToInt(0, 56, true, true),
+  -1,
+);
+assertEqual(
+  toBitArray([0x00, 0xaa, 255, 255, 255, 255, 255]).sliceToInt(
+    0,
+    56,
+    true,
+    false,
+  ),
+  0xaaffffffffff,
+);
+assertEqual(
+  toBitArray([255, 255, 255, 255, 255, 0xaa, 0x00]).sliceToInt(
+    0,
+    56,
+    false,
+    false,
+  ),
+  0xaaffffffffff,
+);
+assertEqual(
+  toBitArray([255, 255, 255, 255, 255, 255, 255]).sliceToInt(
+    0,
+    56,
+    true,
+    false,
+  ),
+  Number(0xfffffffffffffen),
+);
+assertEqual(toBitArray([0xfe, 0x3f]).sliceToInt(4, 12, true, false), 0xe3);
+assertEqual(toBitArray([253, 94]).sliceToInt(3, 11, true, true), -22);
+assertEqual(toBitArray([233, 164]).sliceToInt(3, 15, true, false), 1234);
+assertEqual(toBitArray([250, 72]).sliceToInt(3, 15, false, false), 1234);
+assertEqual(
+  toBitArray([250, 72, 223, 189]).sliceToInt(7, 29, true, false),
+  596983,
+);
+assertEqual(
+  toBitArray([250, 72, 223, 189, 41, 97, 165, 0, 0, 0, 0, 177]).sliceToInt(
+    14,
+    85,
+    false,
+    true,
+  ),
+  70821197049655,
+);
+assertEqual(
+  toBitArray([250, 72, 223, 189, 41, 97, 165, 0, 0, 0, 0, 177]).sliceToInt(
+    14,
+    85,
+    true,
+    true,
+  ),
+  Number(515_906_807_693_217_628_160n),
 );
 
 // Result.isOk
@@ -520,11 +968,12 @@ assertEqual(toList([1, 1]).hasLength(3), false);
 assertEqual([...toList([])], []);
 assertEqual([...toList([1, 2, 3])], [1, 2, 3]);
 
-// BitArray.length
+// BitArray.byteSize
 
-assertEqual(new BitArray(new Uint8Array([])).length, 0);
-assertEqual(new BitArray(new Uint8Array([1, 2])).length, 2);
-assertEqual(new BitArray(new Uint8Array([1, 2, 3, 4])).length, 4);
+assertEqual(new BitArray(new Uint8Array([])).byteSize, 0);
+assertEqual(new BitArray(new Uint8Array([1, 2])).byteSize, 2);
+assertEqual(new BitArray(new Uint8Array([1, 2, 3, 4])).byteSize, 4);
+assertEqual(new BitArray(new Uint8Array([1, 2, 3, 4], 28)).byteSize, 4);
 
 //
 // Division
@@ -579,12 +1028,6 @@ assertEqual(
   new ExampleRecordImpl(1, 2, 3).withFields({ boop: 4, detail: 5, 0: 6 }),
   new ExampleRecordImpl(6, 5, 4),
 );
-
-// Test BitArray can only be constructed from Uint8Array, not ArrayBuffer
-const bs1 = new BitArray(new Uint8Array(new ArrayBuffer(8)));
-assertThrows("Should only construct BitArray from Uint8Array", () => {
-  const bs = new BitArray(new ArrayBuffer(8));
-});
 
 //
 // Summary